Move AccessTokenAuthenticator into the auth module
Also add a try catch around everything
